user_grpc_service/pyproject.toml
[tool.poetry]
name = "user_grpc_service"
version = "0.1.0"
description = ""
authors = ["so1n <qaz6803609@163.com>"]

[tool.poetry.dependencies]
python = "^3.7"
grpcio = "^1.38.1"
grpcio-tools = "^1.38.1"
DBUtils = "^3.0.0"
PyMySQL = "^1.0.2"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
mypy = "^0.910"
pytest = "^6.2.4"
isort = "^5.9.1"
pytest-mock = "^3.6.1"
coverage = "^5.5"
mypy-protobuf = "^2.5"
grpcio-testing = "^1.38.1"
pytest-grpc = "^0.8.0"
pre-commit = "^2.14.0"
flake8 = "^3.9.2"
autoflake = "^1.4"
[tool.black]
line-length = 120
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120

[tool.pytest]
testpaths = "tests"

[tool.coverage.run]
branch = true
source = ["."]
omit = [
    ".venv/*",  "venv/*",
    "tests/*",
    "service/app.py",
    "service/interceptor/base.py",
    "service/interceptor/customer_top.py",
    "service/protos/*",
    "app.py",
]

[tool.coverage.report]
exclude_lines = [
    # Have to re-enable the standard pragma
    "pragma: no cover",

    # Don't complain about missing debug-only code:
    "def __repr__",

    # Don't complain if tests don't hit defensive assertion code:
    "raise AssertionError",
    "raise NotImplementedError",

    # Don't complain if non-runnable code isn't run:
    "if 0:",
    "if __name__ == .__main__.:",

    # type hints
    "if TYPE_CHECKING:",
    "pass",
]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
